requirements.txt
flask
flask-cors
Pillow
numpy
opencv-python-headless
tensorflow
